Stringify Type in Actions
