+++ client/src/app/actionTypes.js
@@ -1,4 +1,5 @@
 const exports = {
+  CANCEL_NAVIGATION: 'CANCEL_NAVIGATION',
   CANCEL_MODAL: 'CANCEL_MODAL',
 };
