+++ client/src/league/actions.js
@@ -1,10 +1,10 @@
 import axios from 'axios';
-import { FETCH_LEAGUES } from '../actions';
+import { FETCH_LEAGUES } from './actionTypes';
 
 export function fetchLeagues() {
   const request = axios.get('/leagues');
   return {
-    type: FETCH_LEAGUES,
+    type: 'FETCH_LEAGUES',
     payload: request,
   };
 }
