+++ client/src/player/actionTypes.js
@@ -1,0 +1,7 @@
+const exports = {
+  GET_ONE_PLAYER_MODAL: 'GET_ONE_PLAYER_MODAL',
+  FILTER_PLAYERS: 'FILTER_PLAYERS',
+  REQUEST_ALL_PLAYERS: 'REQUEST_ALL_PLAYERS',
+};
+
+export default exports;
