+++ client/src/reducers/reducer_leagues.js
@@ -1,10 +1,10 @@
-import { FETCH_LEAGUES } from '../actions/index';
+import { FETCH_LEAGUES } from '../league/actionTypes';
 
 const INITIAL_STATE = [];
 
 export default function (state = INITIAL_STATE, action) {
   switch (action.type) {
-    case FETCH_LEAGUES:
+    case 'FETCH_LEAGUES':
       return action.payload.data;
     default:
       return state;
