+++ client/src/reducers/reducer_players.js
@@ -1,14 +1,15 @@
-import { REQUEST_ALL_PLAYERS, FILTER_PLAYERS, FETCH_SPECIFIC_PLAYERS } from '../actions/index';
+import { REQUEST_ALL_PLAYERS, FILTER_PLAYERS } from '../player/actionTypes';
+import { FETCH_SPECIFIC_PLAYERS } from '../compare/actionTypes';
 
 const INITIAL_STATE = [];
 
 export default function (state = INITIAL_STATE, action) {
   switch (action.type) {
-    case REQUEST_ALL_PLAYERS:
+    case 'REQUEST_ALL_PLAYERS':
       return action.payload.data;
-    case FILTER_PLAYERS:
+    case 'FILTER_PLAYERS':
       return action.payload.data[0];
-    case FETCH_SPECIFIC_PLAYERS:
+    case 'FETCH_SPECIFIC_PLAYERS':
       return action.payload.data[0];
     default:
       return state;
