+++ client/src/routes.js
@@ -2,14 +2,13 @@ import React from 'react';
 
 import { Route, IndexRoute } from 'react-router';
 
-import App from './App';
+import App from './app/App';
 import Compare from './compare';
 import StatisticsOverview from './component/StatisticsOverview.jsx';
 import Schedule from './schedule';
 import TeamView from './yahooTeam';
 import PlayerView from './player';
 import LeagueOverview from './league';
-import YahooProfileImage from './containers/YahooProfileImage';
 
 
 export default (
@@ -18,7 +17,6 @@ export default (
     <Route path="compare" component={Compare} />
     <Route path="schedule" component={Schedule} />
     <Route path="player" component={PlayerView} />
-    <Route path="YahooProfile" component={YahooProfileImage} />
     <Route path="LeagueOverview" component={LeagueOverview} />
     <Route path="Statistics" component={StatisticsOverview} />
     <Route path="TeamView/:league_key" component={TeamView} />
